# Integrate mimalloc allocator