--- a/platform/android/detect.py
+++ b/platform/android/detect.py
@@ -80,7 +80,7 @@ def get_flags():
     return {
         "arch": "arm64",
         "target": "template_debug",
-        "supported": ["mono"],
+        "supported": ["mimalloc", "mono"],
     }