--- a/platform/ios/detect.py
+++ b/platform/ios/detect.py
@@ -50,7 +50,7 @@ def get_flags():
         "target": "template_debug",
         "use_volk": False,
         "metal": True,
-        "supported": ["metal", "mono"],
+        "supported": ["mimalloc", "metal", "mono"],
         "builtin_pcre2_with_jit": False,
     }