--- a/platform/web/detect.py
+++ b/platform/web/detect.py
@@ -79,6 +79,7 @@ def get_flags():
         "target": "template_debug",
         "builtin_pcre2_with_jit": False,
         "vulkan": False,
+        "supported": ["mimalloc"],
         # Embree is heavy and requires too much memory (GH-70621).
         "module_raycast_enabled": False,
         # Use -Os to prioritize optimizing for reduced file size. This is