I noticed #99427, an integration of mimalloc.

Since I already had an integration in a private fork, I've decided to upstream it so we can decide which one is the preferred way.

Main highlight is that this is low-level integration: as such it also hooks into allocations made by 3rd-party code (TLS storage, C/C++ standard library containers, 3rd-party libraries, etc.).

Disclaimer:
This has been thoroughly tested downstream on Windows and certain non-standard platforms. This PR would need some work to add the low-level integration to all the other platforms. On Mac it may work out-of-the-box, though.