@DeeJayLSP 2024-12-14
If it helps somehow, I tried an approach of mimalloc a while ago: https://github.com/DeeJayLSP/godot/commit/eae97092fdf84bd1a2f17ed2cd2160b7909dde96

However, getting it to work properly was beyond my knowledge.


One thing that I did manage to do however, was linking with a system-provided mimalloc for Linux packages. You just have to insert in `platform/linuxbsd/detect.py` a `builtin_mimalloc` check that parses `pkg-config mimalloc --cflags --libs` if false.