@YYF233333 2025-01-09
I’d like to ping this PR as it hasn’t been updated for a long time. Is there any issue holding it up? Hope it can be merged since the average 10% performance improvement should help alleviate many editor issues.