@RandomShaper 2025-01-14
Mac and .NET still have issues. I'd like to work on that once I have some certainty that this is wanted. Now it looks as though. As soon as I can, I'll finish it.