@YYF233333 2025-01-14
Repeat all benchmarks in #99427 with this PR (on top of latest master):

<details>
<summary>Results</summary>

### Instantiate with Many Properties
from #97829 
Run 10 loops calculate average time:
- master
Average: tt1: 314.500000ms, tt2: 434.200000ms, diff: 1.380830
- mimalloc
Average: tt1: 280.500000ms, tt2: 373.900000ms, diff: 1.333078

Time reduced: 10.8% and 13.9%

### Node Duplicate
from #91230
- master: 2135 ms
- mimalloc: 1650 ms

Time reduced: 22.7%

### Loading RichTextLabel with lots of Text
from #96475
command: `time .\godot.windows.editor.x86_64.exe --path ".\RichTextLabel\" --quit`

- master: 11.18 s
- mimalloc: 7.56 s

Time reduced: 32.4%

### Button and Label Instantiate
from #71182
increase button count to 10000
- master 2980 +- 30ms
- mimaloc 2560 +- 10ms

Time reduced: 14.1%

### Empty Project Open
command: `time .\godot.windows.editor.x86_64.exe --editor --path "./emptyProject" --quit`

- master: 7.43 s
- mimalloc: 6.76 s

Time reduced: 9.0%

All tests repeated at least 3 times and ensure results are stable.
</details>

This PR performs even better than mine as it override malloc for thirdparty code so more codes benifit from faster memory operations. Hope this can attract some attention on this :)