@beicause 2025-03-01
I investigated the reason for mono build crash. It's due to the `RTLD_DEEPBIND` flag in `dlopen` in `OS::get_singleton()->open_dynamic_library`, which is called by `load_hostfxr`
https://github.com/godotengine/godot/blob/15ff450680a40391aabbffde0a57ead2cd84db56/drivers/unix/os_unix.cpp#L90-L91
https://github.com/godotengine/godot/blob/15ff450680a40391aabbffde0a57ead2cd84db56/drivers/unix/os_unix.cpp#L895

The `RTLD_DEEPBIND` flag makes the dynamic library prioritize using its own defined allocation symbols instead of mimalloc and causes a crash. After removing the flag in `os_unix.cpp`, I can make the mono version work in Linux.

Note in Android, `RTLD_DEEPBIND` is not set in `dlopen`, so mono version works too.