@dsnopek 2025-03-01
> It's due to the `RTLD_DEEPBIND` flag in `dlopen` in `OS::get_singleton()->open_dynamic_library`, which is called by `load_hostfxr`

We need `RTLD_DEEPBIND` for GDExtensions - see https://github.com/godotengine/godot/pull/82973

However, it's possible to tell if `open_dynamic_library()` is being used to load a GDExtension or not, by checking if the 3rd argument (`GDExtensionData *p_data`) is set or is `nullptr`. So, one option is to only use it when loading GDExtensions.