@beicause 2025-03-01
Or we could use the approach like #99427, or include `mimalloc-override.h` in `memory.cpp`. (Not sure about the exact reason, but this method won't cause problems in mono build).