@DeeJayLSP 2025-03-01
> Or we could use the approach like #99427, or include `mimalloc-override.h` in `memory.cpp`. (Not sure about the exact reason, but this method won't cause problems in mono build).

Including `mimalloc-override.h` in memory.cpp creates a broken build, as some code in thirdparty wouldn't be using mimalloc.

Moreover, this way doesn't go along well with ThreadSanitizer.