@beicause 2025-03-01
I haven't tested it yet. I'm wondering that since this PR breaks Mono, will it also break GDExtension?