@dsnopek 2025-03-01
> I haven't tested it yet. I'm wondering that since this PR breaks Mono, will it also break GDExtension?

It's worth testing, but it _shouldn't_ break it.

GDExtension already has a mechanism to call into Godot and use Godot's allocator for Godot objects. And if the GDExtension needs to use its own allocator for its own objects, it should continue to do so because of the `RTLD_DEEPBIND`.