@Veradictus 2025-06-24
> > However, it's possible to tell if `open_dynamic_library()` is being used to load a GDExtension or not, by checking if the 3rd argument (`GDExtensionData *p_data`) is set or is `nullptr`. So, one option is to only use it when loading GDExtensions.
> 
> I tried and this works for linux mono build (CI passed, haven't test if it works with real projects), see https://github.com/YYF233333/godot/tree/godot_mimalloc.

I've tested this PR in production with my game and all is fine. Performance gain is noticeable.