@YYF233333 2025-09-26
Is there a plan to advance this PR during the 4.6 development cycle? Since the memory allocator has a relatively wide impact, I think it would be better to get it merged before dev1, so that there will be sufficient time for testing.