@DeeJayLSP 2025-09-26
I managed to get a macOS unit test that doesn't crash at start, but still SIGSEGVs at one test: https://github.com/DeeJayLSP/godot/actions/runs/18046970726/job/51360131908

It seems you need to define `MI_OSX_INTERPOSE` and `MI_OSX_ZONE`.