@clayjohn 2025-09-26
Just a heads up. We do not intend to merge this anytime soon. We have discussed this at Core meetings a few times and every time we come to the conclusion that we need to re-evaluate the performance impact _after_ we have improved our memory allocation usage patterns. Currently we allocate way more memory at run time than we should. Both in terms of frequency of allocations and size of allocations. 

We are working towards having near zero allocations per frame. In which case the potential performance benefit of mimalloc would be reduced significantly. If we achieve zero allocations per frame, then the benefit of mimalloc would be restricted to load times and may not have much impact.