@DeeJayLSP 2025-09-26
> We are working towards having near zero allocations per frame.

As an example, I decided to repeat the Dictionary stress test from the [Dictionary reserve PR](https://github.com/godotengine/godot/pull/110709#issue-3436141378). The benchmark itself should have reduced allocations with that PR applied.

Before and after #110709:
```
# Before
time untyped: 7083ms
time typed: 7257ms

# After
time untyped: 6227ms
time typed: 6552ms
```

Before and after #110709 but with `mimalloc` on `LD_PRELOAD`:
```
# Before
time untyped: 4858ms
time typed: 5390ms

# After
time untyped: 4420ms
time typed: 4810ms
```

The improvement of that PR went somewhere from a 12% to 10% with mimalloc on the equation.