@RandomShaper 2025-10-01
Given the current plans to work on reducing allocations first (which makes full sense to me) and therefore it's not prioritary for me to keep trying to polish this PR, I'm marking it as draft. Anyone that wants to take it to the next level, either for own use, experiments or for the potential future reassessment can feel free to take over.