@Saul2022 2026-02-19
> We are working towards having near zero allocations per frame. In which case the potential performance benefit of mimalloc would be reduced significantly. If we achieve zero allocations per frame, then the benefit of mimalloc would be restricted to load times and may not have much impact.

If i may ask, has there been progress on this stuff?